+++ backend/src/dogs/dogs.entity.ts
@@ -2,7 +2,7 @@ import { Column, Entity, JoinColumn, ManyToOne, OneToOne, PrimaryGeneratedColumn
 import { Breed } from '../breed/breed.entity';
 import { DogWalkDay } from '../dog-walk-day/dog-walk-day.entity';
 import { TodayWalkTime } from '../today-walk-time/today-walk-time.entity';
-import { GENDER, Gender } from './types/dog.type';
+import { GENDER, Gender } from './types/gender.type';
 
 @Entity('dogs')
 export class Dogs {
