+++ backend/src/dogs/dtos/create-dog.dto.ts
@@ -1,5 +1,5 @@
 import { IsBoolean, IsEnum, IsNotEmpty, IsOptional, IsPositive, IsString } from 'class-validator';
-import { GENDER, Gender } from '../types/dog.type';
+import { GENDER, Gender } from '../types/gender.type';
 
 export class CreateDogDto {
     @IsString()
@@ -24,36 +24,6 @@ export class CreateDogDto {
 
     @IsPositive()
     @IsNotEmpty()
-    weight: number;
-
-    @IsString()
-    @IsOptional()
-    profilePhotoUrl: string | null;
-}
-
-export class UpdateDogDto {
-    @IsString()
-    @IsOptional()
-    name: string;
-
-    @IsString()
-    @IsOptional()
-    breed: string;
-
-    @IsEnum(GENDER)
-    @IsOptional()
-    gender: Gender;
-
-    @IsBoolean()
-    @IsOptional()
-    isNeutered: boolean;
-
-    @IsString()
-    @IsOptional()
-    birth: string | null;
-
-    @IsPositive()
-    @IsOptional()
     weight: number;
 
     @IsString()
