+++ backend/src/dogs/dtos/update-dog.dto.ts
@@ -1,0 +1,32 @@
+import { IsBoolean, IsEnum, IsOptional, IsPositive, IsString } from 'class-validator';
+import { GENDER, Gender } from '../types/gender.type';
+
+export class UpdateDogDto {
+    @IsString()
+    @IsOptional()
+    name: string;
+
+    @IsString()
+    @IsOptional()
+    breed: string;
+
+    @IsEnum(GENDER)
+    @IsOptional()
+    gender: Gender;
+
+    @IsBoolean()
+    @IsOptional()
+    isNeutered: boolean;
+
+    @IsString()
+    @IsOptional()
+    birth: string | null;
+
+    @IsPositive()
+    @IsOptional()
+    weight: number;
+
+    @IsString()
+    @IsOptional()
+    profilePhotoUrl: string | null;
+}
