+++ backend/src/dogs/types/dog-profile.type.ts
@@ -1,0 +1,12 @@
+import { Gender } from './gender.type';
+
+export interface DogProfile {
+    id: number;
+    name: string;
+    breed: string;
+    gender: Gender;
+    isNeutered: boolean;
+    birth: string | null;
+    weight: number;
+    profilePhotoUrl: string | null;
+}
