+++ backend/src/dogs/types/dog-summary.type.ts
@@ -1,0 +1,5 @@
+export interface DogSummary {
+    id: number;
+    name: string;
+    profilePhotoUrl: string | null;
+}
