+++ backend/src/dogs/types/gender.type.ts
@@ -1,0 +1,6 @@
+export const GENDER = {
+    Male: 'MALE',
+    Female: 'FEMALE',
+} as const;
+
+export type Gender = (typeof GENDER)[keyof typeof GENDER];
